+++ Sources/Dub/API/Models/Click.swift
@@ -1,0 +1,12 @@
+//
+//  Click.swift
+//  Dub
+//
+//  Created by Ian MacCallum on 9/18/25.
+//
+
+import Foundation
+
+public struct Click: Codable, Sendable {
+    public let id: String
+}
